Adding page tracking
Currently switching pages only swaps it for any page specific changes required. Adding in to track the current page once swapped so that things like adding a table of contents or other pages mid document are feasible. Currently it will always add a new page to the end of the document which if pages are swapped is not ideal